fix: prettier + extract isTypeNode
